+++ publishable/database/seeds/PermissionsTableSeeder.php
@@ -27,15 +27,9 @@ class PermissionsTableSeeder extends Seeder
 
         Permission::generateFor('menus');
 
-        Permission::generateFor('pages');
-
         Permission::generateFor('roles');
 
         Permission::generateFor('users');
-
-        Permission::generateFor('posts');
-
-        Permission::generateFor('categories');
 
         Permission::generateFor('settings');
     }
