+++ publishable/database/seeds/VoyagerDummyDatabaseSeeder.php
@@ -7,7 +7,7 @@ class VoyagerDummyDatabaseSeeder extends Seeder
 {
     use Seedable;
 
-    protected $seedersPath = __DIR__.'/';
+    protected $seedersPath;
 
     /**
      * Run the database seeds.
@@ -16,6 +16,7 @@ class VoyagerDummyDatabaseSeeder extends Seeder
      */
     public function run()
     {
+        $this->seedersPath = database_path('seeds').'/';
         $this->seed('CategoriesTableSeeder');
         $this->seed('UsersTableSeeder');
         $this->seed('PostsTableSeeder');
