+++ src/VoyagerServiceProvider.php
@@ -2,8 +2,6 @@
 
 namespace TCG\Voyager;
 
-use Arrilot\Widgets\Facade as Widget;
-use Arrilot\Widgets\ServiceProvider as WidgetServiceProvider;
 use Illuminate\Contracts\Events\Dispatcher;
 use Illuminate\Database\Eloquent\Collection;
 use Illuminate\Foundation\AliasLoader;
@@ -24,6 +22,7 @@ use TCG\Voyager\Models\Setting;
 use TCG\Voyager\Policies\BasePolicy;
 use TCG\Voyager\Policies\MenuItemPolicy;
 use TCG\Voyager\Policies\SettingPolicy;
+use TCG\Voyager\Providers\VoyagerDummyServiceProvider;
 use TCG\Voyager\Providers\VoyagerEventServiceProvider;
 use TCG\Voyager\Translator\Collection as TranslatorCollection;
 
@@ -46,7 +45,7 @@ class VoyagerServiceProvider extends ServiceProvider
     {
         $this->app->register(VoyagerEventServiceProvider::class);
         $this->app->register(ImageServiceProvider::class);
-        $this->app->register(WidgetServiceProvider::class);
+        $this->app->register(VoyagerDummyServiceProvider::class);
         $this->app->register(VoyagerHooksServiceProvider::class);
         $this->app->register(DoctrineSupportServiceProvider::class);
 
@@ -61,7 +60,6 @@ class VoyagerServiceProvider extends ServiceProvider
 
         $this->registerAlertComponents();
         $this->registerFormFields();
-        $this->registerWidgets();
 
         $this->registerConfigs();
 
@@ -222,19 +220,6 @@ class VoyagerServiceProvider extends ServiceProvider
     }
 
     /**
-     * Register widget.
-     */
-    protected function registerWidgets()
-    {
-        $default_widgets = ['TCG\\Voyager\\Widgets\\UserDimmer', 'TCG\\Voyager\\Widgets\\PostDimmer', 'TCG\\Voyager\\Widgets\\PageDimmer'];
-        $widgets = config('voyager.dashboard.widgets', $default_widgets);
-
-        foreach ($widgets as $widget) {
-            Widget::group('voyager::dimmers')->addWidget($widget);
-        }
-    }
-
-    /**
      * Register the publishable files.
      */
     private function registerPublishableResources()
@@ -248,14 +233,8 @@ class VoyagerServiceProvider extends ServiceProvider
             'seeds' => [
                 "{$publishablePath}/database/seeds/" => database_path('seeds'),
             ],
-            'demo_content' => [
-                "{$publishablePath}/demo_content/" => storage_path('app/public'),
-            ],
             'config' => [
                 "{$publishablePath}/config/voyager.php" => config_path('voyager.php'),
-            ],
-            'migrations' => [
-                "{$publishablePath}/database/migrations/" => database_path('migrations'),
             ],
 
         ];
